+++ 06-06-2025/QuizupAPI/Models/DTOs/QuestionAddRequestDTO.cs
@@ -10,7 +10,7 @@ namespace QuizupAPI.Models.DTOs.Question
         public List<string> Options { get; set; } = new List<string>();
 
         [Required]
-        public int QuizId { get; set; }
+        public long QuizId { get; set; }
 
         [Required]
         public string CorrectAnswer { get; set; } = string.Empty;
